Реализовал CRUD для версий заявки, добавление пустого черновика версии при создании заявки

+++ src/main/java/ru/jsms/backend/articles/service/OfferArticleService.java
@@ -4,12 +4,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.jsms.backend.articles.dto.request.CreateOfferArticleRequest;
 import ru.jsms.backend.articles.dto.request.EditOfferArticleRequest;
-import ru.jsms.backend.common.dto.PageParam;
 import ru.jsms.backend.articles.dto.response.OfferArticleResponse;
-import ru.jsms.backend.common.dto.PageDto;
 import ru.jsms.backend.articles.entity.OfferArticle;
-import ru.jsms.backend.articles.enums.OfferArticleStatus;
+import ru.jsms.backend.articles.enums.PublishingStatus;
 import ru.jsms.backend.articles.repository.OfferArticleRepository;
+import ru.jsms.backend.common.dto.PageDto;
+import ru.jsms.backend.common.dto.PageParam;
 import ru.jsms.backend.profile.service.AuthService;
 
 import static ru.jsms.backend.articles.enums.ArticleExceptionCode.ACCESS_DENIED;
@@ -22,6 +22,7 @@ public class OfferArticleService {
 
     private final OfferArticleRepository offerArticleRepository;
     private final AuthService authService;
+    private final OfferArticleVersionService versionService;
 
     public PageDto<OfferArticleResponse> getOfferArticles(PageParam pageParam) {
         final Long userId = (Long) authService.getAuthInfo().getPrincipal();
@@ -31,16 +32,21 @@ public class OfferArticleService {
 
     public OfferArticleResponse createOfferArticle(CreateOfferArticleRequest request) {
         final Long userId = (Long) authService.getAuthInfo().getPrincipal();
-        return convertToResponse(offerArticleRepository.save(OfferArticle.builder()
+        OfferArticle offerArticle = offerArticleRepository.save(OfferArticle.builder()
                 .name(request.getName())
                 .ownerId(userId)
                 .build()
-        ));
+        );
+        versionService.createDefaultVersion(offerArticle.getId());
+        return convertToResponse(offerArticle);
     }
 
     public void deleteOfferArticle(Long id) {
         offerArticleRepository.findById(id).ifPresent(o -> {
             validateAccess(o);
+            if (o.getStatus() != PublishingStatus.DRAFT) {
+                throw EDIT_DENIED.getException();
+            }
             offerArticleRepository.delete(o);
         });
     }
@@ -48,6 +54,9 @@ public class OfferArticleService {
     public OfferArticleResponse editOfferArticle(Long id, EditOfferArticleRequest request) {
         OfferArticle offerArticle = offerArticleRepository.findById(id).orElseThrow(ARTICLE_NOT_FOUND.getException());
         validateAccess(offerArticle);
+        if (offerArticle.getStatus() != PublishingStatus.DRAFT) {
+            throw EDIT_DENIED.getException();
+        }
         offerArticle.setName(request.getName());
         return convertToResponse(offerArticleRepository.save(offerArticle));
     }
@@ -56,9 +65,6 @@ public class OfferArticleService {
         final Long userId = (Long) authService.getAuthInfo().getPrincipal();
         if (!offerArticle.getOwnerId().equals(userId)) {
             throw ACCESS_DENIED.getException();
-        }
-        if (offerArticle.getStatus() != OfferArticleStatus.DRAFT) {
-            throw EDIT_DENIED.getException();
         }
     }
